Move network-device code into its own file.
